+++ cartoframes/context.py
@@ -7,7 +7,6 @@ import os
 import random
 import sys
 import collections
-import binascii as ba
 from warnings import warn
 
 import requests
@@ -1685,65 +1684,3 @@ class CartoContext(object):
                                                      str_value[-50:])
             print('{key}: {value}'.format(key=key,
                                           value=str_value))
-
-
-# TODO: move all of the below to the utils module
-def _add_encoded_geom(df, geom_col):
-    """Add encoded geometry to DataFrame"""
-    # None if not a GeoDataFrame
-    is_geopandas = getattr(df, '_geometry_column_name', None)
-    if is_geopandas is None and geom_col is None:
-        warn('`encode_geom` works best with Geopandas '
-             '(http://geopandas.org/) and/or shapely '
-             '(https://pypi.python.org/pypi/Shapely).')
-        geom_col = 'geometry' if 'geometry' in df.columns else None
-        if geom_col is None:
-            raise KeyError('Geometries were requested to be encoded '
-                           'but a geometry column was not found in the '
-                           'DataFrame.'.format(geom_col=geom_col))
-    elif is_geopandas and geom_col:
-        warn('Geometry column of the input DataFrame does not '
-             'match the geometry column supplied. Using user-supplied '
-             'column...\n'
-             '\tGeopandas geometry column: {}\n'
-             '\tSupplied `geom_col`: {}'.format(is_geopandas,
-                                                geom_col))
-    elif is_geopandas and geom_col is None:
-        geom_col = is_geopandas
-    # updates in place
-    df['the_geom'] = df[geom_col].apply(_encode_geom)
-    return None
-
-
-def _encode_decode_decorator(func):
-    """decorator for encoding and decoding geoms"""
-    def wrapper(*args):
-        """error catching"""
-        try:
-            processed_geom = func(*args)
-            return processed_geom
-        except ImportError as err:
-            raise ImportError('The Python package `shapely` needs to be '
-                              'installed to encode or decode geometries. '
-                              '({})'.format(err))
-    return wrapper
-
-
-@_encode_decode_decorator
-def _encode_geom(geom):
-    """Encode geometries into hex-encoded wkb
-    """
-    from shapely import wkb
-    if geom:
-        return ba.hexlify(wkb.dumps(geom)).decode()
-    return None
-
-
-@_encode_decode_decorator
-def _decode_geom(ewkb):
-    """Decode encoded wkb into a shapely geometry
-    """
-    from shapely import wkb
-    if ewkb:
-        return wkb.loads(ba.unhexlify(ewkb))
-    return None
